requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1705494584803/work
argcomplete==3.4.0
argh==0.31.2
astunparse @ file:///home/conda/feedstock_root/build_artifacts/astunparse_1610696312422/work
biopython==1.83
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1695989787169/work
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1718025014955/work/certifi
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.15.4
flatbuffers @ file:///home/conda/feedstock_root/build_artifacts/python-flatbuffers_1711466727397/work
fonttools==4.53.0
fsspec==2024.6.0
gast @ file:///home/conda/feedstock_root/build_artifacts/gast_1688368721366/work
gffutils==0.13
h5py @ file:///home/conda/feedstock_root/build_artifacts/h5py_1717664771270/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1713279365350/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1719361860083/work
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
keras @ file:///home/conda/feedstock_root/build_artifacts/keras_1714206291078/work
kiwisolver==1.4.5
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1710435156458/work
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1686175045316/work
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1706899945820/work
matplotlib==3.9.0
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1704317613764/work
ml-dtypes @ file:///home/conda/feedstock_root/build_artifacts/ml_dtypes_1704727673911/work
mpmath==1.3.0
namex @ file:///home/conda/feedstock_root/build_artifacts/namex_1713169736451/work
networkx==3.2.1
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opt-einsum @ file:///home/conda/feedstock_root/build_artifacts/opt_einsum_1696448916724/work
optree @ file:///home/conda/feedstock_root/build_artifacts/optree_1711402432724/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
protobuf==4.25.3
psutil==6.0.0
pyfaidx==0.8.1.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyparsing==3.1.2
pysam==0.22.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1717057054362/work
rich @ file:///home/conda/feedstock_root/build_artifacts/rich-split_1709150387247/work/dist
scikit-learn==1.5.0
scipy==1.13.1
sentry-sdk==2.6.0
setproctitle==1.3.3
simplejson==3.19.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
spliceai-toolkit==0.0.1
sympy==1.12.1
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1708285739699/work/tensorboard-2.16.2-py3-none-any.whl#sha256=9f2b4e7dad86667615c0e5cd072f1ea8403fc032a299f0072d6f74855775cc45
tensorboard-data-server @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-data-server_1695425348895/work/tensorboard_data_server-0.7.0-py3-none-manylinux2014_x86_64.whl#sha256=c3327dfe496d955cacc10e4977436bc235dece0c0640c862ede133f2fd842b28
tensorflow @ file:///home/conda/feedstock_root/build_artifacts/tensorflow-split_1716512380594/work/tensorflow_pkg/tensorflow-2.16.1-cp39-cp39-linux_x86_64.whl#sha256=8a821c1c15506e3f375683d9b598d574f392924e81096f58d5610317911ead79
tensorflow_estimator @ file:///home/conda/feedstock_root/build_artifacts/tensorflow-split_1716512380594/work/tensorflow-estimator/wheel_dir/tensorflow_estimator-2.15.0-py2.py3-none-any.whl#sha256=b92387d2dab9e3665a8ce20ff9b975a5acd931b680d2d91cc050c1d32b697afc
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1704357939450/work
threadpoolctl==3.5.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
triton==2.3.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1718653059220/work
wandb==0.17.3
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1715000201436/work
wrapt @ file:///home/conda/feedstock_root/build_artifacts/wrapt_1699532794105/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1718013267051/work
